+++ src/components/cell.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {StyleSheet, View, TextInput} from 'react-native';
+import { StyleSheet, View, TextInput } from 'react-native';
 
 export const Header = props => {
   return <View style={styles.header}>{props.children}</View>;
@@ -11,18 +11,6 @@ export const CellTitleContainer = props => {
 
 export const CellContainer = props => {
   return <View style={styles.cell}>{props.children}</View>;
-};
-
-export const VBox = props => {
-  return <View style={styles.vbox}>{props.children}</View>;
-};
-
-export const VSeparator = props => {
-  if (props.half) {
-    return <View style={styles.vSeparatorHalf}>{props.children}</View>;
-  } else {
-    return <View style={styles.vSeparator}>{props.children}</View>;
-  }
 };
 
 export const TextInputCell = props => {
@@ -59,15 +47,6 @@ const styles = StyleSheet.create({
     borderColor: '#e3e3e3',
     backgroundColor: '#f2f2f2',
     borderWidth: 0.5,
-  },
-  vbox: {
-    marginHorizontal: 16,
-  },
-  vSeparator: {
-    height: 32,
-  },
-  vSeparatorHalf: {
-    height: 16,
   },
   ListSeparator: {
     height: 1,
